[REFACTORING] : ChatList RepositoryImple addUser()구현
