+++ app/src/main/java/com/erionna/eternalreturninfo/ui/repository/ChatListRepository.kt
@@ -1,7 +1,9 @@
 package com.erionna.eternalreturninfo.ui.repository
 
+import com.erionna.eternalreturninfo.model.ERModel
+
 interface ChatListRepository {
-    fun adduser()
+    fun addUser(item: ERModel?) : List<ERModel>
     fun currentList()
     fun clearList()
     fun modifyItemForCallBack()
