+++ app/src/main/java/com/erionna/eternalreturninfo/ui/viewmodel/ChatListViewModel.kt
@@ -5,11 +5,13 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.erionna.eternalreturninfo.data.repository.ChatListRepositoryImpl
 import com.erionna.eternalreturninfo.model.ERModel
+import com.erionna.eternalreturninfo.ui.repository.ChatListRepository
 import java.util.concurrent.atomic.AtomicLong
 
 class ChatListViewModel(
-    private val idGenerate: AtomicLong
+    val repository: ChatListRepository
 ) : ViewModel() {
     private val _list: MutableLiveData<List<ERModel>> = MutableLiveData()
     val list: LiveData<List<ERModel>> get() = _list
@@ -17,48 +19,7 @@ class ChatListViewModel(
     fun addUser(
         item: ERModel?
     ) {
-        if (item == null) {
-            return
-        }
-
-        val sb = StringBuilder()
-        var time = ""
-        if (item.time != "") {
-            sb.append(item.time)
-            time = sb.substring(0, 13)
-        } else {
-            time = ""
-        }
-
-        val currentList = list.value.orEmpty().toMutableList()
-        var gotCha: Boolean? = null
-
-        if (currentList.size > 0) {
-
-            for (i in 0 until currentList.size) {
-                if (currentList[i].name == item.name) {
-                    gotCha = true
-                }
-            }
-            if (gotCha != true) {
-                _list.value = currentList.apply {
-                    add(
-                        item.copy(
-                            time = time
-                        )
-                    )
-                }
-            }
-
-        } else {
-            _list.value = currentList.apply {
-                add(
-                    item.copy(
-                        time = time
-                    )
-                )
-            }
-        }
+        _list.value = repository.addUser(item)
     }
 
     fun currentList(): List<ERModel> {
@@ -128,11 +89,13 @@ class ChatListViewModel(
 }
 
 class ChatListViewModelFactory : ViewModelProvider.Factory {
-    private val idGenerate = AtomicLong(1L)
+    private val repository: ChatListRepository = ChatListRepositoryImpl(
+        AtomicLong(1L)
+    )
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(ChatListViewModel::class.java)) {
-            return ChatListViewModel(idGenerate) as T
+            return ChatListViewModel(repository) as T
         } else {
             throw IllegalArgumentException("not found correct viewModel")
         }
